Create credentials stored in Cleartext
A vulnerability has been identified in the Asus RT-N12+ B1, specifically related to how the router stores its sensitive information (passwords) in cleartext in its database. This exposes users to potential security risks, as cleartext passwords are easily accessible/readable as they are not encrypted. Immediate action is recommended to address this critical flaw to enhance the router's security.