Create CSV Injection
The CSV Injection vulnerability in the Asus RT-N12+ router enables users to inject and execute arbitrary commands or formulas within the context of a CSV. 
When a user opens a spreadsheet, an Office-based application (such as Excel on Windows, LibreOffice, and OpenOffice), it will parse cells containing meta-characters as spreadsheet formulas. Applications that are vulnerable to formula injection allow untrusted user input to appear in a generated spreadsheet, triggering arbitrary command execution when the spreadsheet is opened. Attackers can use the malicious payload to invoke dangerous programs, using native system tools such as “cmd.exe” and “powershell.exe” to remotely compromise an individual who downloaded and opened the malicious CSV file in one of the spreadsheets software mentioned above. Generally, the user needs to accept the warning details for the command to be executed.